Refactor to prepare for the second pipeline.
