+++ example-trace-messages/cons/__init__.py
@@ -1,4 +1,5 @@
 from .cons import mkenv
-from .prompt_to_md import prompt_to_md
+from .pipelines import prompt_to_md
+from .demo_run import build_plan_writing_trace
 
-__all__ = ["mkenv", "prompt_to_md"]
+__all__ = ["mkenv", "prompt_to_md", "build_plan_writing_trace"]
