+++ example-trace-messages/traces/basic_hello/020_state.json
