+++ example-trace-messages/traces/basic_hello/030_state.json
