+++ example-trace-messages/traces/basic_hello/05_stdin.txt
