+++ example-trace-messages/traces/basic_hello/060_state.json
