+++ example-trace-messages/traces/basic_hello/070_state.json
